Archiv/products/slack_bot/requirements.txt
slack-bolt>=1.18.0
slack-sdk>=3.21.0
Flask>=2.3.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
requests>=2.31.0